Apply suggestion word on empty lines
Fix the bug of not applying the attribute line in a selected word
when the attribute line is previous applied in an empty line.
For example, when we applied a list in an empty line, and after it we
select a word in a list of suggestion box without type anything.