Fix Latin chars when suggestion is selected
When user types non-Latin chars and selects a suggestion that does have
them, the resulting text might not be correct. Ex: suggestion is
"árvore", user typed "arv"; resulting text would be "arvore", which is
incorrect.

This is the first implementation to fix the issue, it still needs to be
optimized.