Avoid scrolling page down when user presses Ctrl+space
Some machines/browsers are configured to scroll page down when space is
pressed and browser is not on edit mode. When Ctrl is pressed with
space, Etherpad leaves edit mode and thus Ctrl+space was opening
suggestion box but also scrolling the page. This commit avoid page to
scroll when this shortcut is pressed.