Store line number that made suggestion box be opened
This is useful for other plugins that extend ep_autocomp, if they need a
reference of why suggestions were being displayed. We store line number
instead of the DOM element or the whole line (from rep.lines) because
those values are modified by Etherpad every time a change is made on the
pad. If we stored any of those values it would be impossible* to compare
them with the actual state of the pad.

*: or at least I did not find any way to do that.