[feat] Allow autocomp obj to be imported by other files using `require`
Besides exposing the `autocomp` object on a variable on `window`, expose
it also on the `static/js/index` module, so other plugins that depend on
`ep_autocomp` can explicitly let Etherpad core know that.

This might be especially useful when optimizing plugin loading using
https://github.com/storytouch/ep_webpack.

Other improvements on this commit:

- unify client-side hooks on a single file, to follow the standard of
other Etherpad plugins;
- use our own copy of `sendkeys`, to avoid depending on a file we don't
have control over (the file on tests of Etherpad core). This also avoids
breaking this plugin behavior when Etherpad test files are not exposed
in production instances.